backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pypdf
pikepdf
img2pdf
pillow
